divide reducer on two parts
