+++ src/components/main/main.jsx
@@ -6,9 +6,11 @@ import LocationsList from "../locations-list/location-list.jsx";
 import SortOptions from "../sort-options/sort-options.jsx";
 import {connect} from "react-redux";
 import {CITIES, SortType} from "../../utils.js";
-import {ActionCreator} from "../../reducer/reducer.js";
+import {ActionCreator} from "../../reducer/main/main.js";
 import {createSelector} from "reselect";
 import MainEmpty from "../main-empty/main-emty.jsx";
+import {getMarker} from "../../reducer/main/selectors.js";
+import {selectOffers, getActiveCity} from "../../reducer/offers/selectors.js";
 
 
 const Main = (props) => {
@@ -18,6 +20,8 @@ const Main = (props) => {
     // eslint-disable-next-line react/prop-types
     marker, handleOfferHover,
   } = props;
+
+  console.log(props)
 
   const advertsCount = offers.length;
 
@@ -101,33 +105,33 @@ Main.propTypes = {
   sortType: PropTypes.string,
 };
 
-function selectOffers(state) {
-  return state.offers;
-}
+// function selectOffers(state) {
+//   return state.offers;
+// }
 
-function getSortedOffers(offers, sortType) {
-  switch (sortType) {
-    case SortType.PRICE_TO_LOW:
-      return offers.slice().sort((a, b) => b.price - a.price);
-    case SortType.PRICE_TO_HIGH:
-      return offers.slice().sort((a, b) => a.price - b.price);
-    case SortType.TOP_RATED:
-      return offers.slice().sort((a, b) => b.rating - a.rating);
-  }
-  return offers;
-}
-
-const sortOffersBySortType = createSelector([
-  selectOffers,
-  (state) => state.sortType
-], (offers, sortType) => getSortedOffers(offers, sortType)
-);
+// function getSortedOffers(offers, sortType) {
+//   switch (sortType) {
+//     case SortType.PRICE_TO_LOW:
+//       return offers.slice().sort((a, b) => b.price - a.price);
+//     case SortType.PRICE_TO_HIGH:
+//       return offers.slice().sort((a, b) => a.price - b.price);
+//     case SortType.TOP_RATED:
+//       return offers.slice().sort((a, b) => b.rating - a.rating);
+//   }
+//   return offers;
+// }
+//
+// const sortOffersBySortType = createSelector([
+//   selectOffers,
+//   (state) => state.MAIN.sortType
+// ], (offers, sortType) => getSortedOffers(offers, sortType)
+// );
 
 const mapStateToProps = (state) => {
   return {
-    activeCity: state.activeCity,
-    offers: sortOffersBySortType(state),
-    marker: state.marker
+    activeCity: getActiveCity(state),
+    offers: selectOffers(state),
+    marker: getMarker(state)
   };
 };
