+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import App from "./components/app/app.jsx";
 import {createStore} from 'redux';
 import {Provider} from 'react-redux';
-import {reducer} from "./reducer/reducer.js";
+import reducer from "./reducer/reducer.js";
 
 
 const store = createStore(
