+++ src/reducer/name-space.js
@@ -1,0 +1,6 @@
+const NameSpace = {
+  OFFERS: `OFFERS`,
+  MAIN: `MAIN`,
+};
+
+export default NameSpace;
